// web/src/api/generated/model/operationRunSummarySchemaStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Mixd
 * Personal music metadata hub
 * OpenAPI spec version: 0.10.2.5
 */

export type OperationRunSummarySchemaStatus = typeof OperationRunSummarySchemaStatus[keyof typeof OperationRunSummarySchemaStatus];


export const OperationRunSummarySchemaStatus = {
  running: 'running',
  complete: 'complete',
  partial: 'partial',
  error: 'error',
  cancelled: 'cancelled',
} as const;